Update payloadSizeTracker to use double to avoid overflow

+++ src/api/PayloadSizeTracker.cc
@@ -10,19 +10,19 @@ using namespace Analytical;
 
 PayloadSizeTracker::PayloadSizeTracker(int dims_count) noexcept
     : dims_count(dims_count) {
-  payload_size_sent_through_dim = std::vector<int>(dims_count, 0);
+  payload_size_sent_through_dim = std::vector<double>(dims_count, 0);
 }
 
 void PayloadSizeTracker::addPayloadSize(int payload_size, int dim) noexcept {
   payload_size_sent_through_dim[dim] += payload_size;
 }
 
-int PayloadSizeTracker::payloadSizeThroughDim(int dim) const noexcept {
+double PayloadSizeTracker::payloadSizeThroughDim(int dim) const noexcept {
   return payload_size_sent_through_dim[dim];
 }
 
-int PayloadSizeTracker::totalPayloadSize() const noexcept {
-  auto total_payload = 0;
+double PayloadSizeTracker::totalPayloadSize() const noexcept {
+  auto total_payload = 0.0;
   for (const auto payload : payload_size_sent_through_dim) {
     total_payload += payload;
   }
